Initial version of scenario-based unit tests
